+++ internal/service/platform/connector/k8s_cluster.go
@@ -143,6 +143,11 @@ func ResourceConnectorK8s() *schema.Resource {
 							Type:        schema.TypeString,
 							Required:    true,
 						},
+						"ca_cert_ref": {
+							Description: "Reference to the secret containing the CA certificate for the connector." + secret_ref_text,
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
 					},
 				},
 			},
@@ -418,20 +423,8 @@ func buildConnectorK8s(d *schema.ResourceData) *nextgen.ConnectorInfo {
 			if attr := config["service_account_token_ref"].(string); attr != "" {
 				saConfig.ServiceAccountTokenRef = attr
 			}
-		}
-
-		if attr, ok := d.GetOk("service_account"); ok {
-			config := attr.([]interface{})[0].(map[string]interface{})
-			saConfig := &nextgen.KubernetesServiceAccount{}
-			connector.K8sCluster.Credential.ManualConfig.Auth.Type_ = nextgen.KubernetesAuthTypes.ServiceAccount
-			connector.K8sCluster.Credential.ManualConfig.Auth.ServiceAccount = saConfig
-
-			if attr := config["master_url"].(string); attr != "" {
-				connector.K8sCluster.Credential.ManualConfig.MasterUrl = attr
-			}
-
-			if attr := config["service_account_token_ref"].(string); attr != "" {
-				saConfig.ServiceAccountTokenRef = attr
+			if attr := config["ca_cert_ref"].(string); attr != "" {
+				saConfig.CaCertRef = attr
 			}
 		}
 
@@ -517,6 +510,7 @@ func readConnectorK8s(d *schema.ResourceData, connector *nextgen.ConnectorInfo) 
 				{
 					"master_url":                connector.K8sCluster.Credential.ManualConfig.MasterUrl,
 					"service_account_token_ref": auth.ServiceAccount.ServiceAccountTokenRef,
+					"ca_cert_ref":               auth.ServiceAccount.CaCertRef,
 				},
 			})
 			d.Set("delegate_selectors", connector.K8sCluster.DelegateSelectors)